+++ Hashtable.h
@@ -21,14 +21,14 @@ using namespace std;
 
 class Hash_Member{
 	private :
-		struct Item<int>* item;
+		struct Item<double>* item;
 		G* g_m;
 	public:
-		struct Item<int> * get_item() { return item; }
+		struct Item<double> * get_item() { return item; }
 
-		Hash_Member(G* g_f,Item<int>* it);
+		Hash_Member(G* g_f,Item<double>* it);
 		~Hash_Member();
-		struct Item<int>* Combine(vector<int>& c,G *g_h,double& temp_dist);
+		struct Item<double>* Combine(vector<double>& c,G *g_h,double& temp_dist);
 };
 
 class Hashtable{
@@ -38,9 +38,9 @@ class Hashtable{
 	public :
 		Hashtable(int& k,int& d);
 		~Hashtable();
-		void Insert_Hashtable(struct Item<int>* item,uint32_t& t_size);
-		struct Item<int>* NN_Hashtable(vector<int>& c,uint32_t& t_size,double& dist);
-		void Range_Hashtable(vector<int>& c,uint32_t& t_size,double& R,vector<struct Item <int>*>& range);
+		void Insert_Hashtable(struct Item<double>* item,uint32_t& t_size);
+		struct Item<double>* NN_Hashtable(vector<double>& c,uint32_t& t_size,double& dist);
+		void Range_Hashtable(vector<double>& c,uint32_t& t_size,double& R,vector<struct Item <double>*>& range);
 };
 
 #endif
