+++ Search.h
@@ -17,7 +17,7 @@
 
 using namespace std;
 
-void Range(vector<int>& c,unordered_map<int,vector <Item<int>*>>& h,int& metric,double& R,int& M,int& probes,Cosine_G* cg,G* g,vector<struct Item <int>*>& range);
-struct Item<int>* NN(vector<int>& c,unordered_map<int,vector <Item<int>*>>& h,int& metric,int& M,int& probes,Cosine_G* cg,G* g,double& min_dist);
+void Range(vector<double>& c,unordered_map<int,vector <Item<double>*>>& h,int& metric,double& R,int& M,int& probes,Cosine_G* cg,G* g,vector<struct Item <double>*>& range);
+struct Item<double>* NN(vector<double>& c,unordered_map<int,vector <Item<double>*>>& h,int& metric,int& M,int& probes,Cosine_G* cg,G* g,double& min_dist);
 
 #endif
